(rendering) Extract level-circle/icon-in-ring engine into LevelCircleMixin
Move the Chart.js doughnut rendering engine (_renderLevelCircle,
_rebuildCharts, _buildLevelRingConfig, icon-in-ring and color helpers,
plus the chart lifecycle hooks) into a reusable mixin so a forthcoming
badge element can share it without duplicating any code. No behavior
change to the card: the card keeps its forecast-subscription logic in
updated()/disconnectedCallback() and chains into the mixin via super.
